WIP: permissions

+++ src/images/images.service.ts
@@ -1,6 +1,7 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CloudinaryService } from 'src/cloudinary/cloudinary.service';
+import { User } from '@prisma/client';
 
 @Injectable()
 export class ImagesService {
@@ -50,23 +51,55 @@ export class ImagesService {
     return reportPhotos;
   }
 
-  async createImageReport(id: string, image: Express.Multer.File) {
-    const uploadedImage = await this.cloudinaryService.uploadFile(image, id);
-    const newReportImage = await this.prisma.reportPhoto.create({
-      data: {
-        url: uploadedImage.secure_url,
-        reportId: id
-      },
-      select: {
-        url: true,
-        id: true
-      }
-    })
+  async createImageReport(id: string, image: Express.Multer.File, user: User) {
 
-    return newReportImage;
+    const report = await this.prisma.report.findUnique({
+      where: { id }
+    });
+
+    if (!report) {
+      throw new BadRequestException('El reporte no existe');
+    }
+
+    if(user.id !== report.studentId) {
+      throw new BadRequestException('No tienes permisos para crear esta imagen' );
+    }
+
+    try {
+      const uploadedImage = await this.cloudinaryService.uploadFile(image, id);
+      const newReportImage = await this.prisma.reportPhoto.create({
+        data: {
+          url: uploadedImage.secure_url,
+          reportId: id
+        },
+        select: {
+          url: true,
+          id: true
+        }
+      })
+  
+      return newReportImage;
+      
+    } catch (error) {
+      throw new BadRequestException('Error al crear la imagen');
+    }
   }
 
-  async deleteImage(id: string) {
+  async deleteImage(id: string, user: User) {
+
+    const report = await this.prisma.report.findUnique({
+      where: { id }
+    });
+
+    if (!report) {
+      throw new BadRequestException('El reporte no existe');
+    }
+
+    if(user.id !== report.studentId) {
+      throw new BadRequestException('No tienes permisos para eliminar esta imagen' );
+    }
+    
+
     try {
       const image = await this.prisma.reportPhoto.findUnique({
         where: { id }
@@ -95,7 +128,21 @@ export class ImagesService {
     }
   }
 
-  deleteFolderWithImages(folderName: string) {
+  async deleteFolderWithImages(folderName: string, user: User) {
+
+    const report = await this.prisma.report.findUnique({
+      where: { id: folderName }
+    });
+
+    if (!report) {
+      throw new BadRequestException('El reporte no existe');
+    }
+
+    if(user.role !== 'admin') {
+      throw new BadRequestException('No tienes permisos para borrar la carpeta' );
+    }
+    
+
     return this.cloudinaryService.deleteFolderWithImages(folderName); 
   }
